backend/requirements-minimal.txt
# Minimal requirements for backend API
fastapi==0.104.1
uvicorn[standard]==0.24.0
click>=8.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database essentials
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0

# Redis basics
redis==5.0.1
aioredis==2.0.1

# Core ML/Data
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
tensorflow==2.15.0
keras==2.15.0
ta==0.11.0

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.25.2

# Crypto exchange
ccxt==4.2.25

# WebSocket
websockets==12.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0
psutil==5.9.6

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0